Refactor: Extract country_code_to_emoji_flag into Country_Flag trait
Move the duplicate country_code_to_emoji_flag() method from both Feedback
and Feedback_Field classes into a shared Country_Flag trait to eliminate
code duplication.
